# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Meridian API
  version: 0.18.0
  description: Same-origin REST contract for owner authentication, journal evidence, Triage, internal tasks/reminders, provider-independent Today, goals, local planning, execution evidence, The Weekly, local knowledge ingestion, worker health, and Microsoft consent.
paths:
...
  /api/knowledge/sources/{sourceId}:
    get:
      operationId: getKnowledgeSource
      summary: Read one source, its immutable revisions, and exact-span claims.
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/KnowledgeSourceIdPath' }]
      responses:
        '200': { $ref: '#/components/responses/KnowledgeSourceDetail' }
        '401': { $ref: '#/components/responses/SessionInvalid' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-meridian-session
      description: Production cookie name; local development omits the __Host prefix.
  parameters:
...
    KnowledgeSourceIdPath:
      name: sourceId
      in: path
      required: true
      schema: { type: string, format: uuid }
  schemas:
...
    KnowledgeSourceResponseV1:
      type: object
      additionalProperties: true
      required:
        [
          id,
          title,
          sourceClass,
          reviewStatus,
          correctionStatus,
          deletionRequestedAt,
          createdAt,
          updatedAt,
          version,
        ]
      properties:
        id: { type: string, format: uuid }
        title: { type: string, minLength: 1, maxLength: 500 }
        sourceClass:
          type: string
          enum:
            [
              systematic_review_or_meta_analysis,
              randomised_trial,
              controlled_non_randomised_study,
              observational_study,
              mechanistic_or_laboratory_study,
              clinical_or_professional_guideline,
              narrative_review,
              expert_commentary,
              book_or_chapter,
              podcast_or_transcript,
              personal_notes,
              unknown,
            ]
        reviewStatus:
          type: string
          enum:
            [
              unreviewed,
              processing,
              reviewed,
              reference_only,
              rejected,
              superseded,
            ]
        correctionStatus:
          type: string
          enum:
            [unknown, none_known, corrected, retracted, expression_of_concern]
        deletionRequestedAt:
          type: [string, 'null']
          format: date-time
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        version: { type: integer, minimum: 1 }
...
    KnowledgeSourceDetailResponseV1:
      type: object
      additionalProperties: false
      required: [source, revisions, claims]
      properties:
        source: { $ref: '#/components/schemas/KnowledgeSourceResponseV1' }
        revisions:
          type: array
          items:
            type: object
            additionalProperties: true
            required:
              [
                id,
                sourceId,
                revisionNumber,
                originalContentHash,
                parsedText,
                extractionQuality,
                processingClass,
                chunkCount,
              ]
            properties:
              id: { type: string, format: uuid }
              sourceId: { type: string, format: uuid }
              revisionNumber: { type: integer, minimum: 1 }
              originalContentHash: { type: string, pattern: '^[a-f0-9]{64}$' }
              parsedText: { type: string }
              extractionQuality:
                {
                  type: string,
                  enum: [complete, partial, ocr_required, failed],
                }
              processingClass:
                { type: string, enum: [standard, sensitive, private] }
              chunkCount: { type: integer, minimum: 0 }
        claims:
          type: array
          items:
            type: object
            additionalProperties: true
            required:
              [
                id,
                sourceId,
                claimText,
                epistemicStatus,
                reviewStatus,
                citations,
                version,
              ]
            properties:
              id: { type: string, format: uuid }
              sourceId: { type: string, format: uuid }
              claimText: { type: string, minLength: 1, maxLength: 4000 }
              epistemicStatus: { type: string, const: reported_by_source }
              reviewStatus:
                {
                  type: string,
                  enum: [candidate, reviewed, rejected, superseded],
                }
              citations:
                type: array
                minItems: 1
                items:
                  type: object
                  additionalProperties: true
                  required:
                    [
                      sourceRevisionId,
                      sourceSpanStart,
                      sourceSpanEnd,
                      quotedTextHash,
                    ]
                  properties:
                    sourceRevisionId: { type: string, format: uuid }
                    sourceSpanStart: { type: integer, minimum: 0 }
                    sourceSpanEnd: { type: integer, minimum: 1 }
                    quotedTextHash: { type: string, pattern: '^[a-f0-9]{64}$' }
              version: { type: integer, minimum: 1 }
    ErrorResponseV1:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: string
          enum:
            [
              AUTHENTICATION_FAILED,
              CONFLICT,
              CSRF_INVALID,
              INTERNAL_ERROR,
              INTEGRATION_UNAVAILABLE,
              INVALID_AUTHORITY,
              NOT_FOUND,
              PROCESSING_CLASS_VIOLATION,
              PROHIBITED_ACTION,
              RATE_LIMITED,
              SESSION_INVALID,
              VALIDATION_FAILED,
            ]
  responses:
...
    SessionInvalid:
      description: Missing, expired, revoked, or invalid session.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponseV1' }
...
    KnowledgeSourceDetail:
      description: Owner-scoped local source, immutable revisions, and exact-span claims.
      content:
        application/json:
          schema:
            { $ref: '#/components/schemas/KnowledgeSourceDetailResponseV1' }
    NotFound:
      description: The owner-scoped resource does not exist.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponseV1' }
...